Listados

+++ EquiposTecnicosSN.Web/Controllers/RepuestosController.cs
@@ -1,14 +1,12 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Data;
+﻿using System.Data;
 using System.Data.Entity;
 using System.Linq;
 using System.Threading.Tasks;
 using System.Net;
-using System.Web;
 using System.Web.Mvc;
 using EquiposTecnicosSN.Entities.Mantenimiento;
 using EquiposTecnicosSN.Web.DataContexts;
+using PagedList;
 
 namespace EquiposTecnicosSN.Web.Controllers
 {
@@ -49,10 +47,22 @@ namespace EquiposTecnicosSN.Web.Controllers
         }
 
         // GET: Repuestos
-        public async Task<ActionResult> Index()
+        public ActionResult Index(string searchNombre = null, string searchCodigo = null, int proveedorId = 0, int page = 1)
         {
-            var repuestos = db.Repuestos.Include(r => r.Proveedor);
-            return View(await repuestos.ToListAsync());
+            var listPage = db.Repuestos.Include(r => r.Proveedor)
+                .Where(r => searchNombre == null || r.Nombre.Contains(searchNombre))
+                .Where(r => searchCodigo == null || r.Codigo.Contains(searchCodigo))
+                .Where(r => proveedorId == 0 || r.ProveedorId == proveedorId)
+                .OrderBy(r => r.Nombre)
+                .ToPagedList(page, 10);
+
+            if (Request.IsAjaxRequest())
+            {
+                return PartialView("_RepuestosList", listPage);
+            }
+            
+            ViewBag.proveedorId = new SelectList(db.Proveedores, "ProveedorId", "Nombre");
+            return View(listPage);
         }
 
         // GET: Repuestos/Details/5
@@ -129,13 +139,13 @@ namespace EquiposTecnicosSN.Web.Controllers
         }
 
         // GET: Repuestos/Delete/5
-        public async Task<ActionResult> Delete(int? id)
+        public ActionResult Delete(int? id)
         {
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            Repuesto repuesto = await db.Repuestos.FindAsync(id);
+            Repuesto repuesto = db.Repuestos.Find(id);
             if (repuesto == null)
             {
                 return HttpNotFound();
@@ -146,12 +156,45 @@ namespace EquiposTecnicosSN.Web.Controllers
         // POST: Repuestos/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public async Task<ActionResult> DeleteConfirmed(int id)
+        public ActionResult DeleteConfirmed(int id)
         {
-            Repuesto repuesto = await db.Repuestos.FindAsync(id);
-            db.Repuestos.Remove(repuesto);
-            await db.SaveChangesAsync();
-            return RedirectToAction("Index");
+            var stockCount = db.StockRepuestos
+                .Where(sr => sr.RepuestoId == id)
+                .Count();
+
+            var solicitudesCount = db.SolicitudesRepuestosServicios
+                .Where(sr => sr.RepuestoId == id)
+                .Count();
+
+            var equiposCount = db.InformacionesComerciales
+                .Where(ic => ic.ProveedorId == id)
+                .Count();
+
+            var repuesto = db.Repuestos.Find(id);
+
+            if (equiposCount == 0 && solicitudesCount == 0 && stockCount == 0)
+            {
+                db.Repuestos.Remove(repuesto);
+                db.SaveChanges();
+                return RedirectToAction("Index");
+            }
+
+            if (equiposCount != 0)
+            {
+                ModelState.AddModelError("", "El repuesto no puede eliminarse ya que posee equipos asociados.");
+            }
+
+            if (solicitudesCount != 0)
+            {
+                ModelState.AddModelError("", "El repuesto no puede eliminarse ya que posee pedidos asociados.");
+            }
+
+            if (stockCount != 0)
+            {
+                ModelState.AddModelError("", "El repuesto no puede eliminarse ya que posee un stock asociado.");
+            }
+
+            return View(repuesto);
         }
 
         protected override void Dispose(bool disposing)
